+++ src/main/java/hexlet/code/specification/TaskFilter.java
@@ -14,7 +14,7 @@ public class TaskFilter {
                 .and(withLabelId(params.getLabelId()));
     }
 
-    private Specification<Task> withAssigneeId(Integer assigneeId) {
+    private Specification<Task> withAssigneeId(Long assigneeId) {
         return (root, query, cb) -> assigneeId == null ? cb.conjunction() : cb.equal(root.get("assignee")
                 .get("id"), assigneeId);
     }
@@ -29,7 +29,7 @@ public class TaskFilter {
                 .equal(root.get("taskStatus").get("slug"), status);
     }
 
-    private Specification<Task> withLabelId(Integer labelId) {
+    private Specification<Task> withLabelId(Long labelId) {
         return (root, query, criteriaBuilder) -> labelId == null ? criteriaBuilder.conjunction() : criteriaBuilder
                 .equal(root.get("labels").get("id"), labelId);
     }
